refactor(tokens): split sign-user-token to allow safe usage in Vercel Edge Function
