+++ src/index.ts
@@ -7,3 +7,6 @@ export * from "./common/userTokens";
 export * from "./common/exceptions";
 
 export { Knock } from "./knock";
+
+// Export separately to allow using in a lightweight runtime environment, e.g. Edge Function, without pulling in full Knock library.
+export { signUserToken } from "./sign-user-token";
